Pipelines CLI Improvements (#8780)
* Rename `wrangler pipelines show` to `wrangler pipelines get`

* Remove `--enable-http|worker-binding` in favor of `--source <sources...>` format

Replace `--enable-worker-binding` and `--enable-http` with `--source worker`
and `--source http` (or `--source http worker` for both).

This also refactors how these bindings get added to the pipeline
configuration, opting for a handler pattern instead of multiple if
statements.

* Remove `--file-template` and `--partition-template` flags from `wrangler pipelines create|update`

* Prettify `wrangler pipelines get <pipeline>` output ✨

* Clarify defaults, minimums, and maximums for pipelines commands

We specifically do not use yargs `default` field as these values are
purely informational and the API is the ultimate decider of what the
default value will be at a given time.

* Show pipelines command in wrangler help

* update snapshots

* Fix updating CORS origins for pipelines

* Fix upper limit on batch-max-rows

---------

Co-authored-by: Peter Bacon Darwin <pbacondarwin@cloudflare.com>