fix(vite-plugin): Support `__` in additional module IDs (#9322)
* fix(vite-plugin): Anchor ADDITIONAL_MODULE_PATTERN regex to end of string

 ## Problem

There's a regex used to detect if certain files (wasm, .bin, .txt, .html) are being imported at runtime. The current regex looks like this:

```ts
const ADDITIONAL_MODULE_PATTERN = `__CLOUDFLARE_MODULE__(${ADDITIONAL_MODULE_TYPES.join("|")})__(.*?)__`;
```

This means if there are files with `__`s in them, those `__`s will get matched first, rather than the final `__` in the module pattern being matched on.

 ## Solution
Add a `$` to the end of the regex so, to anchor to the end of the string.

 ## Background context
`pnpm` stores files for packages in locations like this (note the `__`s), which uncovered this.

```
/path/to/project/node_modules/.pnpm/@prisma+client@6.8.2_prisma@6.8.2_typescript@5.8.3__typescript@5.8.3/node_modules/@prisma/client/runtime/query_engine_bg.sqlite.wasm
```

* Make regex more resilient

* Add test to playground

* Revert fix to demonstrate failing test

* Revert "Revert fix to demonstrate failing test"

This reverts commit d7e57e0fb7ce5b8ec626d3c62165222144ddcf90.

* Revert fix again so we can make sure test failing first

* Revert "Revert fix again so we can make sure test failing first"

This reverts commit 50093d47a636daef151e3f8b482e30fe9c27c999.

* Add changeset