feat(wrangler): add support for FedRAMP High compliance region (#9321)
* feat(wrangler): add support for FedRAMP High compliance region

Now it is possible to target Wrangler at the FedRAMP High compliance region.
There are two ways to signal to Wrangler to run in this mode:

- set `"compliance_region": "fedramp_high"` in a Wrangler configuration
- set `CLOUDFLARE_COMPLIANCE_REGION=fedramp_high` environment variable when running Wrangler

If both are provided then the environment variable overrides the configured value.

When in this mode OAuth authentication is not supported.
It is necessary to authenticate using a Cloudflare API Token acquired from the Cloudflare FedRAMP High dashboard.

Most bindings and commands are supported in this mode.

- Unsupported commands may result in API requests that are not supported - possibly 422 Unprocessable Entity responses.
- Unsupported bindings may work in local dev, as there is no local validation, but will fail at Worker deployment time.

Resolves DEVX-1921.

* update workers.dev domain for fed mode

* error when env var and config don't line up

* fix env-var factory regression from earlier commits in this PR

* ensure that the local runtime controller passes through region to start mixed mode

* Check the logged-in state in the remote runtime controller rather than at the start of `wrangler dev`.

By this point we have the resolved config, so we can user the compliance region
but also it handles the case where a user toggles remote mode on during a dev session.

* Print compliance region in `wrangler whoami`

* fixups

* move ComplianceConfig type to misc-variables

* more fixes

* agh

* tighten ComplianceConfig so that it must be defined

* revert unnecessary change to `usingLocalNamespace`

* fix up the tests

* remove unnecessary empty line