Router worker performs free tier limiting (#9661)
* WC-3584 Refactor routing to asset or user worker

Centralizing some of these pieces will make it easier to add free tier
limiting

* WC-3584 Add eyeball routing config

Provided via a new optional param binding. When provided, it contains
info about free tier limiting for this account.

* WC-3584 Deny requests to the user worker when over free tier limits

Returns an html page (included as direct module, via esbuild) with a 429
status code, mirroring the current free tier limiting behavior. It also
sets a new field in our analytics so we can observe free tier invocation
denials.

This page can evolve as we need it to, this is just my first pass at the
response.

* WC-3584 Add changeset

* WC-3584 Move the limited response page to be inlined

There's no loader configured for html files by default in wrangler it
seems. So rather than mess with loaders or a custom build, inlining this
page is easy enough.

This pages copies what fl returns exactly, and there's a few templated
things to reproduce that. Namely: the hostname in the title of the page,
the ip address of the eyeball, and the current date (with some formatting)