Add support for custom instance types (#10051)
* Add support for custom instance types

Custom instance types could previously only be set through setting
vcpu/memory/disk through the `configuration` block. As we deprecate
`configuration`, move setting vcpu, memory, and disk to the instance
type field of wrangler. By nature, this enforces that custom instance
types are mutually exclusive with the named instance types and will
allow us to remove a lot of the logic that verifies this once
`configuration` is removed.