Wrangler preview urls default to disabled (#10478)
* feat: extract subdomain deploy into its own function

Should make the logic easier to understand.

* fix: config preview_urls defaults to undefined (falsy) instead of true

We want preview_urls to be opt-in, rather than opt-out. For this reason
its default value will be falsy.

* test: fix Wrangler tests with routes

Some of the previous commits that changed defaulting uncovered that most
Wrangler Wrangler tests that had at least one route, were not mocking the
appropriate calls to zones and routes endpoints.

* feat: add Cloudflare-Workers-Script-Api-Date header

Wrangler differs from the Dash and other clients in that it makes a lot
decisions for the user regarding the status of the Worker subdomain. The
header should help with existing and future breaking changes between
Wrangler and the API.

* fix: add warning when subdomain status does not match Wrangler config

Warn the user when the workers.dev or preview URLs status does not match
their Wrangler config file.