feat: Update R2 Data Catalog compaction commands to work on catalog level (remove table/namespace) (#10622)
* [R2 Data Catalog] Update r2 catalog compaction commands to enable/disable catalog

For our intial open-beta, we want customers to have an easier interface
for enabling compaction on their catalogs. This change allows them to
enable once, so any table they create will have compaction enabled.

We will add the ability to control specific tables at a later time.

* [R2 Data Catalog] Enable compaction requires a CLOUDFLARE_API_TOKEN

For compaction to work, we need customers to grant our system permission
to access their R2 Bucket and R2 Data Catalog on their behalf. This
permission grant is made via a token they provide when they enable
compaciton. We encrypt the token and store it, and use it only via our
backend system. Because the token is required for compaction to work, we
make it required.

* Add change set file

* Change the token in 'compaction enable' to be cli arg instead of CLOUDFALRE_API_TOKEN